Change long age

+++ Rider_git_sample1/People.cs
@@ -4,5 +4,5 @@ public class People
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int Age { get; set; }
+    public long Age { get; set; }
 }
